+++ modules/core/dependency/python-ihm/ihm/restraint.py
@@ -249,7 +249,7 @@ class ExperimentalCrossLink(object):
 
 
 class DistanceRestraint(object):
-    """Abstract base class for all distance restraints. These are typically
+    """Base class for all distance restraints. These are typically
        used in a :class:`DerivedDistanceRestraint`.
 
        Do not use this class directly but instead use a derived class
@@ -258,6 +258,7 @@ class DistanceRestraint(object):
        :class:`LowerBoundDistanceRestraint`,
        or :class:`LowerUpperBoundDistanceRestraint`.
     """
+    restraint_type = None
 
     #: The minimum distance allowed for this restraint,
     #: or None if unconstrained